Changes the SA index structure to support more genome sequences (chrom., decoy, ALT...): the CHROM table contains 16-bit values instead of 8-bit values

+++ src/dna/sa_dna_commons.c
@@ -9,7 +9,7 @@ extern sa_genome3_t *global_genome;
 
 int is_cal_in(seed_cal_t *cal) {
   // Y_9911717_9912231_0:1:0_1:0:0_1677c
-  unsigned int chrom;
+  unsigned short int chrom;
   size_t start, end;
   char *chrom_str, *aux, *p1, *p2;
   
@@ -705,7 +705,7 @@ void create_alignments(array_list_t *cal_list, fastq_read_t *read,
 
 void display_suffix_mappings(int strand, size_t r_start, size_t suffix_len, 
 			     size_t low, size_t high, sa_index3_t *sa_index) {
-  unsigned int chrom;
+  unsigned short int chrom;
   size_t r_end, g_start, g_end;
   for (size_t suff = low; suff <= high; suff++) {
     r_end = r_start + suffix_len - 1;
@@ -731,7 +731,7 @@ void print_seed(char *msg, seed_t *s) {
 
 void display_sequence(uint j, sa_index3_t *index, uint len) {
   char *p = &index->genome->S[index->SA[j]];
-  unsigned int chrom = (unsigned int) index->CHROM[j];
+  unsigned short int chrom = (unsigned short int) index->CHROM[j];
   for (int i = 0; i < len; i++) {
     printf("%c", *p);
     p++;
@@ -751,7 +751,7 @@ char *get_subsequence(char *seq, size_t start, size_t len) {
 //--------------------------------------------------------------------
 
 void display_cmp_sequences(fastq_read_t *read, sa_index3_t *sa_index) {
-  unsigned int chrom;
+  unsigned short int chrom;
   size_t pos, strand;
   char *ref, *seq, *chrom_str, *aux, *p1, *p2;
